mcp-notebook/mcp-langchain-adapter/pyproject.toml
[project]
name = "mcp-langchain-adapter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-mcp-adapters>=0.1.10",
    "langchain[openai]>=0.3.27",
    "langgraph>=0.6.7",
    "python-dotenv>=1.1.1",
    "isort>=6.0.1",
    "fastmcp>=2.12.3",
    "httpx>=0.28.1",
    "black>=25.9.0",
]
